Move fence functions into fence.cpp
