+++ src/openrct2/paint/map_element/fence.c
@@ -384,7 +384,3 @@ void fence_paint(uint8 direction, sint32 height, rct_map_element * map_element)
 
     sub_98199C(scrolling_text_setup(stringId, scroll, scrollingMode), 0, 0, 1, 1, 13, height + 8, boundsOffset.x, boundsOffset.y, boundsOffset.z, get_current_rotation());
 }
-
-uint8 fence_get_animation_frame(rct_map_element *fenceElement) {
-	return (fenceElement->properties.fence.animation >> 3) & 0xF;
-}
